FINAL StudentService Refactor
