+++ src/Hogwarts.java
@@ -30,12 +30,5 @@ public abstract class Hogwarts {
 
     public void setMsTransgress(int msTransgress) { this.msTransgress = msTransgress; }
 
-    public static void compareMagic(Hogwarts student1, Hogwarts student2) {
-        if (student1.getMsConjure() > student2.getMsConjure()) {
-            System.out.println(student1.getName() + " обладает бОльшей мощностью магии, чем " + student2.getName() + ".");
-        } else {
-            System.out.println(student2.getName() + " обладает бОльшей мощностью магии, чем " + student1.getName() + ".");
-        }
-    }
 
 }
